Improve tests for the ScreenGroupService

+++ src/main/java/org/worldbridge/development/screenserver/rest/ScreenGroupService.java
@@ -1,6 +1,7 @@
 package org.worldbridge.development.screenserver.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.worldbridge.development.screenserver.dao.DaoException;
 import org.worldbridge.development.screenserver.dao.ScreenGroupDao;
 import org.worldbridge.development.screenserver.domain.DeviceDetails;
 import org.worldbridge.development.screenserver.domain.ScreenGroup;
@@ -16,8 +17,12 @@ import java.util.*;
 @Produces(MediaType.APPLICATION_JSON)
 public class ScreenGroupService {
 
-    @Autowired
     private ScreenGroupDao screenGroupDao;
+
+    @Autowired
+    public void setScreenGroupDao(ScreenGroupDao screenGroupDao) {
+        this.screenGroupDao = screenGroupDao;
+    }
 
     @GET
     public List<ScreenGroup> list() {
@@ -25,25 +30,41 @@ public class ScreenGroupService {
     }
 
     @PUT
-    public Response addScreenGroup(ScreenGroup screenGroup) {
-        screenGroupDao.createScreenGroup(screenGroup.getGroupName());
+    public void addScreenGroup(ScreenGroup screenGroup) {
+        if (screenGroup == null) {
+            throw new BadRequestException();
+        } else if (screenGroup.getDevices().size() != 0) {
+            throw new BadRequestException("Adding a group with devices is not yet supported");
+        }
 
-        return Response.accepted().build();
+        screenGroupDao.createScreenGroup(screenGroup.getGroupName());
     }
 
     @PUT
     @Path("{groupName}/devices")
-    public Response addDeviceToGroup(@PathParam("groupName") String groupName, DeviceDetails deviceDetails) {
-        screenGroupDao.addScreenToScreenGroup(groupName, deviceDetails.getDeviceId());
+    public void addDeviceToGroup(@PathParam("groupName") String groupName, DeviceDetails deviceDetails) {
+        if (deviceDetails == null) {
+            throw new BadRequestException("deviceDetails shouldn't be null");
+        }
 
-        return Response.accepted().build();
+        try {
+            screenGroupDao.addScreenToScreenGroup(groupName, deviceDetails.getDeviceId());
+        } catch (DaoException e) {
+            throw new BadRequestException(e);
+        }
     }
 
     @DELETE
-    @Path("{groupName/devices")
-    public Response removeDeviceFromGroup(@PathParam("groupName") String groupName, DeviceDetails deviceDetails) {
-        screenGroupDao.removeScreenFromScreenGroup(groupName, deviceDetails.getDeviceId());
+    @Path("{groupName}/devices")
+    public void removeDeviceFromGroup(@PathParam("groupName") String groupName, DeviceDetails deviceDetails) {
+        if (deviceDetails == null) {
+            throw new BadRequestException("deviceDetails shouldn't be null");
+        }
 
-        return Response.accepted().build();
+        try {
+            screenGroupDao.removeScreenFromScreenGroup(groupName, deviceDetails.getDeviceId());
+        } catch (DaoException e) {
+            throw new BadRequestException(e);
+        }
     }
 }
